requirements.txt
google-genai
mcp
requests
python-dotenv
fastapi
uvicorn
sqlalchemy
apscheduler
slowapi
